Abstract join community logic into hook
